+++ app.js
@@ -1,9 +1,9 @@
 import { updateDate } from "./src/global.js";
 import { setGridEvents } from "./src/gridView.js";
-import { appendRollingList } from "./src/rollingAnimation.js";
+import { appendRollingList } from "./src/components/rollingAnimation.js";
 import { setListViewEvents } from "./src/listView.js";
-import { setNavigationButton } from "./src/navigationButton.js";
-import { setCategory } from "./src/category.js";
+import { setNavigationButton } from "./src/components/navigationButton.js";
+import { setCategory } from "./src/components/category.js";
 
 document.addEventListener("DOMContentLoaded", function () {
   updateDate();
