+++ src/core/util.js
